+++ tests/unit/Api/ExceptionHandler/RouteNotFoundExceptionHandlerTest.php
@@ -9,37 +9,37 @@
  * file that was distributed with this source code.
  */
 
-namespace Flarum\Tests\Api\ExceptionHandler;
+namespace Flarum\Tests\unit\Api\ExceptionHandler;
 
 use Exception;
-use Flarum\Api\ExceptionHandler\ModelNotFoundExceptionHandler;
-use Flarum\Tests\Test\TestCase;
-use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Flarum\Api\ExceptionHandler\RouteNotFoundExceptionHandler;
+use Flarum\Http\Exception\RouteNotFoundException;
+use PHPUnit\Framework\TestCase;
 
-class ModelNotFoundExceptionHandlerTest extends TestCase
+class RouteNotFoundExceptionHandlerTest extends TestCase
 {
     private $handler;
 
-    public function init()
+    public function setUp()
     {
-        $this->handler = new ModelNotFoundExceptionHandler;
+        $this->handler = new RouteNotFoundExceptionHandler();
     }
 
     public function test_it_handles_recognisable_exceptions()
     {
         $this->assertFalse($this->handler->manages(new Exception));
-        $this->assertTrue($this->handler->manages(new ModelNotFoundException));
+        $this->assertTrue($this->handler->manages(new RouteNotFoundException()));
     }
 
     public function test_managing_exceptions()
     {
-        $response = $this->handler->handle(new ModelNotFoundException);
+        $response = $this->handler->handle(new RouteNotFoundException);
 
         $this->assertEquals(404, $response->getStatus());
         $this->assertEquals([
             [
                 'status' => '404',
-                'code' => 'resource_not_found'
+                'code' => 'route_not_found'
             ]
         ], $response->getErrors());
     }
